chore: Fix Jest warnings for missing theme provider
